backend/requirements.txt
requests
beautifulsoup4
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
click
langgraph
langchain
langchain-community
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-core
fastapi
uvicorn
python-multipart
psycopg2-binary
authlib
pyjwt